cc-wrapper: Include NIX_CFLAGS_LINK in linkType decision
Without this, the cc-wrapper ends up passing -Wl,-dynamic-linker; but if
NIX_CFLAGS_LINK specifies '-static' you end up with a static binary with
a dynamic linker set.

For glibc, (and possibly with other linkers) this does not seem to be
harmful, but for musl with lld it results in a binary which crashes on
startup.

Signed-off-by: Peter Waller <p@pwaller.net>